javadoc OK + pb d'utf8 resolu

+++ src/lsg/buffs/DragonSlayerRing.java
@@ -3,7 +3,7 @@ package lsg.buffs;
 public class DragonSlayerRing extends Ring {
 
 	public DragonSlayerRing() {
-		super("Dragon Slayer Ring", (float)100);
+		super("Dragon Slayer Ring", 100f);
 	}
 	
 }
