+++ src/lsg/buffs/MoonStone.java
@@ -3,7 +3,7 @@ package lsg.buffs;
 public class MoonStone extends Talisman {
 
 	public MoonStone() {
-		super("MoonStone", (float)75);
+		super("MoonStone", 75f);
 	}
 	
 }
