+++ src/lsg/buffs/RingOfDeath.java
@@ -3,7 +3,7 @@ package lsg.buffs;
 public class RingOfDeath extends Ring {
 
 	public RingOfDeath() {
-		super("Ring Of Death", (float)50);
+		super("Ring Of Death", 50f);
 	}
 	
 }
